changed function names and created util component
