+++ src/utils/timeUtils.ts
@@ -1,0 +1,31 @@
+export const readableTimeFormat = (createdAt: string | Date): string => {
+  if (!createdAt) {
+    return 'unknown time';
+  }
+
+  const now = new Date();
+  const then = new Date(createdAt);
+
+  if (isNaN(then.getTime())) {
+    return 'unknown time';
+  }
+
+  const diffInSeconds = Math.floor((now.getTime() - then.getTime()) / 1000);
+
+  if (diffInSeconds < 60) {
+    return 'just now';
+  }
+
+  const diffInMinutes = Math.floor(diffInSeconds / 60);
+  if (diffInMinutes < 60) {
+    return `${diffInMinutes} minute${diffInMinutes > 1 ? 's' : ''} ago`;
+  }
+
+  const diffInHours = Math.floor(diffInMinutes / 60);
+  if (diffInHours < 24) {
+    return `${diffInHours} hour${diffInHours > 1 ? 's' : ''} ago`;
+  }
+
+  const diffInDays = Math.floor(diffInHours / 24);
+  return `${diffInDays} day${diffInDays > 1 ? 's' : ''} ago`;
+};
